requirements.txt
aiohttp==3.9.1 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==4.2.0 ; python_version >= "3.9" and python_version < "4.0"
asn1crypto==1.5.1 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.9" and python_version < "4.0"
cachecontrol==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.3.2 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
cloud-sql-python-connector[pg8000]==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
cryptography==41.0.7 ; python_version >= "3.9" and python_version < "4.0"
Deprecated==1.2.14 ; python_version >= "3.9" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.9" and python_version < "4.0"
fastapi==0.108.0 ; python_version >= "3.9" and python_version < "4.0"
firebase-admin
frozenlist==1.4.1 ; python_version >= "3.9" and python_version < "4.0"
future==0.18.3 ; python_version >= "3.9" and python_version < "4.0"
gax-google-logging-v2==0.8.3 ; python_version >= "3.9" and python_version < "4.0"
gax-google-pubsub-v1==0.8.3 ; python_version >= "3.9" and python_version < "4.0"
gcloud==0.18.3 ; python_version >= "3.9" and python_version < "4.0"
google-api-core==2.16.1 ; python_version >= "3.9" and python_version < "4.0"
google-api-python-client==2.116.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.26.2 ; python_version >= "3.9" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-firestore==2.14.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-secret-manager==2.18.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-storage==2.14.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-gax==0.12.5 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.7.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.62.0 ; python_version >= "3.9" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-iam-v1==0.13.0 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-logging-v2==0.8.1 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-pubsub-v1==0.8.1 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.60.1 ; python_version >= "3.9" and python_version < "4.0"
grpcio-status==1.60.1 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.9" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
jiter==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.9" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.9" and python_version < "4.0"
jwcrypto==1.5.1 ; python_version >= "3.9" and python_version < "4.0"
msgpack==1.0.7 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
oauth2client==4.1.3 ; python_version >= "3.9" and python_version < "4.0"
openai==1.46.0 ; python_version >= "3.9" and python_version < "4.0"
openapi==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
pg8000==1.30.4 ; python_version >= "3.9" and python_version < "4.0"
ply==3.8 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.25.2 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pycryptodome==3.20.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.5.3 ; python_version >= "3.9" and python_version < "4.0"
pydantic_core==2.14.6 ; python_version >= "3.9" and python_version < "4.0"
PyJWT==2.8.0 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.9" and python_version < "4.0"
Pyrebase4==4.7.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-jwt==4.1.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.9" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.29.0 ; python_version >= "3.9" and python_version < "4.0"
requests-toolbelt==0.10.1 ; python_version >= "3.9" and python_version < "4.0"
rpds-py==0.20.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4.0"
scramp==1.4.4 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==2.0.25 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.32.0.post1 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.66.5 ; python_version >= "3.9" and python_version < "4.0"
typing_extensions==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.18 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.25.0 ; python_version >= "3.9" and python_version < "4.0"
validator==0.7.1 ; python_version >= "3.9" and python_version < "4.0"
validators==0.34.0 ; python_version >= "3.9" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-pubsub
opencv-python
Pillow
